requirements.txt
google-cloud-firestore
selenium
python-slugify
psutil
requests
google-cloud-storage